// src/session/schema.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * This interface was referenced by `undefined`'s JSON-Schema
 * via the `definition` "APIKeysResponse".
 */
export interface APIKeysResponse {
  stripeKey: string;
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema
 * via the `definition` "Auth0OAuthResponse".
 */
export interface Auth0OAuthResponse {
  newAccount: boolean;
  existingEmail: string;
  sessionKey?: string;
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema
 * via the `definition` "BoxAcquireRequest".
 */
export interface BoxAcquireRequest {
  key: string;
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema
 * via the `definition` "BoxAcquireResponse".
 */
export interface BoxAcquireResponse {
  data: string;
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema
 * via the `definition` "BoxProvideRequest".
 */
export interface BoxProvideRequest {
  key: string;
  data: string;
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema
 * via the `definition` "BoxProvideResponse".
 */
export interface BoxProvideResponse {}
/**
 * This interface was referenced by `undefined`'s JSON-Schema
 * via the `definition` "Session".
 */
export interface Session {
  id: string;
  created: string;
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema
 * via the `definition` "SessionsResponse".
 */
export interface SessionsResponse {
  sessions: {
    id: string;
    created: string;
  }[];
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema
 * via the `definition` "WhoamiResponse".
 */
export interface WhoamiResponse {
  sessionAge: number;
  accountId: string;
  email: string;
  firstName: string;
  lastName: string;
  created: number;
  publicKey: string;
  encSymmetricKey: string;
  encPrivateKey: string;
  saltEnc: string;
  isPaymentRequired: boolean;
  isTrialing: boolean;
  isVerified: boolean;
  isAdmin: boolean;
  trialEnd: string;
  planName: string;
  planId: string;
  canManageTeams: boolean;
  maxTeamMembers: number;
  passwordSet: boolean;
  canUnsetPassword: boolean;
  unsubscribedFromEmails: boolean;
}